Convert context1SubthoughtCreated to Typescript and fix logic.
